requirements.txt
flickrapi
pillow
numpy
requests
torch
torchvision
